+++ include/Finch/FinchPasses.td
@@ -57,9 +57,40 @@ def FinchLoopletRun: Pass<"finch-looplet-run"> {
 def FinchLoopletSequence: Pass<"finch-looplet-sequence"> {
   let summary = "Compiler Pass for Sequence Looplets";
   let description = [{
-    Compiler Pass for handling sequence looplets
+    Lowers finch.access of finch.sequence inside scf.for
+    
+    `finch.sequence crd, body1, body2` can be thought as a function
+
+    f(x) = body1 (if x < crd)
+           body2 (if x >= crd)
+
+    if we iterate f(x) over [st, en),
+    f(x) = body1 (if x in [-inf,crd) ∩ [st, en))
+           body2 (if x in [crd,+inf) ∩ [st, en))
+
+    ```mlir
+    %0 = finch.run %cst_0 : (f32) -> (!finch.looplet)
+    %1 = finch.run %cst_1 : (f32) -> (!finch.looplet)
+    %c3 = arith.constant 3 : index
+    %2 = finch.sequence %c3, %0, %1 : (index, !finch.looplet, !finch.looplet) -> (!finch.looplet)
+    scf.for %idx = %st to %en step %c1 {
+        %3 = finch.access %2, %idx : f32
+    }
     ```
-    to be filled.
+    to
+
+    ```mlir 
+    %0 = finch.run %cst_0 : (f32) -> (!finch.looplet)
+    %1 = finch.run %cst_1 : (f32) -> (!finch.looplet)
+    %c3 = arith.constant 3 : index
+    %2 = arith.minui %en, %c3 : index
+    %3 = arith.maxui %st, %c3 : index
+    scf.for %idx = %st to %2 step %c1 {
+      %5 = finch.access %0, %idx : f32
+    }
+    scf.for %arg2 = %3 to %en step %c1 {
+      %5 = finch.access %1, %idx : f32
+    }
 
     ```
   }];
