+++ lib/Finch/FinchPasses.cpp
@@ -25,7 +25,6 @@
 namespace mlir::finch {
 #define GEN_PASS_DEF_FINCHSIMPLIFIER
 #define GEN_PASS_DEF_FINCHINSTANTIATE
-#define GEN_PASS_DEF_FINCHLOOPLETSEQUENCE
 #define GEN_PASS_DEF_FINCHLOOPLETSTEPPER
 #define GEN_PASS_DEF_FINCHLOOPLETLOOKUP
 #include "Finch/FinchPasses.h.inc"
@@ -225,110 +224,6 @@ public:
       return success();
     }
 
-    return failure();
-  }
-};
-
-
-class FinchLoopletSequenceRewriter : public OpRewritePattern<scf::ForOp> {
-public:
-  using OpRewritePattern<scf::ForOp>::OpRewritePattern;
-
-  LogicalResult matchAndRewrite(scf::ForOp forOp,
-                                PatternRewriter &rewriter) const final {
-    auto loopIndex = forOp.getInductionVar();
-    
-    OpBuilder builder(forOp);
-    Location loc = forOp.getLoc();
-    
-    for (auto& accessOp : *forOp.getBody()) {
-      if (!isa<mlir::finch::AccessOp>(accessOp)) 
-        continue;
-      
-      Value accessIndex = accessOp.getOperand(1);
-      if (accessIndex != loopIndex) 
-        continue;
-      
-      auto seqLooplet = dyn_cast<finch::SequenceOp>(
-          accessOp.getOperand(0).getDefiningOp());
-      if (!seqLooplet) 
-        continue;
-         
-      rewriter.setInsertionPoint(forOp);
-      // Main Sequence Rewrite          
-      IRMapping mapper1;
-      IRMapping mapper2;
-      Operation* newForOp1 = rewriter.clone(*forOp, mapper1);
-      Operation* newForOp2 = rewriter.clone(*forOp, mapper2);
-      rewriter.moveOpAfter(newForOp1, forOp);
-      rewriter.moveOpAfter(newForOp2, newForOp1);
-
-      // Replace Access operand with Sequence bodies
-      auto newAccess1 = mapper1.lookupOrDefault(&accessOp);
-      auto newAccess2 = mapper2.lookupOrDefault(&accessOp);
-      auto bodyLooplet1 = seqLooplet.getOperand(1);
-      auto bodyLooplet2 = seqLooplet.getOperand(2);
-      auto newBodyLooplet1 = mapper1.lookupOrDefault(bodyLooplet1);
-      auto newBodyLooplet2 = mapper2.lookupOrDefault(bodyLooplet2);
-      newAccess1->setOperand(0, newBodyLooplet1);
-      newAccess2->setOperand(0, newBodyLooplet2);
-      
-      // Intersection
-      Value loopLb = forOp.getLowerBound();
-      Value loopUb = forOp.getUpperBound();
-
-      // Finch.jl used both closed endpoints,
-      // but Finch.mlir uses [st,en) for now.
-      // This is for aligning syntax with scf.for
-      // scf.for uses [st,en).
-      // TODO: We need to make finch.for
-      //
-      //       firstBodyUb=secondBodyLb
-      //                  v
-      // [---firstBody---)[---secondBody---)
-      Value firstBodyUb = seqLooplet.getOperand(0);
-      Value secondBodyLb = firstBodyUb;
-      if (!firstBodyUb.getType().isIndex()) {
-        firstBodyUb = rewriter.create<arith::IndexCastOp>(
-            loc, rewriter.getIndexType(), firstBodyUb);
-        secondBodyLb = rewriter.create<arith::IndexCastOp>(
-            loc, rewriter.getIndexType(), secondBodyLb);
-      }         
-      
-      // Main intersect
-      Value newFirstLoopUb = rewriter.create<arith::MinUIOp>(
-          loc, loopUb, firstBodyUb);
-      Value newSecondLoopLb = rewriter.create<arith::MaxUIOp>(
-          loc, loopLb, secondBodyLb);
-      cast<scf::ForOp>(newForOp1).setUpperBound(newFirstLoopUb);
-      cast<scf::ForOp>(newForOp2).setLowerBound(newSecondLoopLb);
-
-
-      // Build a chain on scf.for
-      // %0 = tensor.empty()
-      // %1 = scf.for $i = $b0 to %b1 step %c1 iter_args(%v = %0) //forOp
-      // %2 = scf.for $i = $b0 to %b2 step %c1 iter_args(%v = %0) //newForOp1
-      // %3 = scf.for $i = $b2 to %b1 step %c1 iter_args(%v = %0) //newForOp2
-      // return %1
-      //
-      // vvv
-      //
-      // %0 = tensor.empty()
-      // %1 = scf.for $i = $b0 to %b2 step %c1 iter_args(%v = %0) //newForOp1  
-      // %2 = scf.for $i = $b2 to %b1 step %c1 iter_args(%v = %1) //newForOp2
-      // return %2
-
-      // First three are lowerbound, upperbound, and step
-      int numIterArgs = newForOp2->getNumOperands() - 3;
-      if (numIterArgs > 0) {
-        rewriter.replaceAllUsesWith(forOp->getResults(), newForOp2->getResults());
-        newForOp2->setOperands(3, numIterArgs, newForOp1->getResults());
-      }
-
-      rewriter.eraseOp(forOp);
-      
-      return success();
-    }
     return failure();
   }
 };
@@ -600,20 +495,6 @@ public:
     if (failed(applyPatternsGreedily(getOperation(), patternSet))) {
       signalPassFailure();
     }
-  }
-};
-
-class FinchLoopletSequence
-    : public impl::FinchLoopletSequenceBase<FinchLoopletSequence> {
-public:
-  using impl::FinchLoopletSequenceBase<
-      FinchLoopletSequence>::FinchLoopletSequenceBase;
-  void runOnOperation() final {
-    RewritePatternSet patterns(&getContext());
-    patterns.add<FinchLoopletSequenceRewriter>(&getContext()); 
-    FrozenRewritePatternSet patternSet(std::move(patterns));
-    if (failed(applyPatternsGreedily(getOperation(), patternSet)))
-      signalPassFailure();
   }
 };
