+++ lib/Finch/FinchRunPass.cpp
@@ -1,4 +1,4 @@
-//===- FinchLoopletPasses.cpp - Finch Looplet passes -----------------*- C++ -*-===//
+//===- FinchRunPasses.cpp - Finch Run passes -----------------*- C++ -*-===//
 //
 // This file is licensed under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
